feat: add SpecialErroBasePage for error page handling
Introduce SpecialErroBasePage to centralize logic for special error pages based on SyncErrorStates. Refactor DriveAccessDeniedPage to inherit from this new base class, removing duplicated handler and navigation logic. This change improves maintainability and simplifies the creation of future error pages.
