+++ src/gui4/windows/kDrive client/kDrive client/Pages/SpecialErrorsPages/DriveAccessDeniedPage.xaml
@@ -17,7 +17,7 @@
  * along with this program.  If not, see<http://www.gnu.org/licenses/> .
  -->
 
-<Page x:Class="Infomaniak.kDrive.Pages.DriveAccessDeniedPage"
+<local:SpecialErroBasePage x:Class="Infomaniak.kDrive.Pages.DriveAccessDeniedPage"
       xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
       xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
       xmlns:local="using:Infomaniak.kDrive.Pages"
@@ -26,9 +26,9 @@
       xmlns:controls="using:Infomaniak.kDrive.CustomControls"
       xmlns:converters="using:Infomaniak.kDrive.Converters"
       mc:Ignorable="d">
-    <Page.Resources>
+    <local:SpecialErroBasePage.Resources>
         <converters:ColorToBrushConverter x:Key="ColorToBrushConverter" />
-    </Page.Resources>
+    </local:SpecialErroBasePage.Resources>
     <Border Background="{ThemeResource LayerFillColorDefault}"
             BorderThickness="1"
             BorderBrush="{ThemeResource CardStrokeColorDefault}"
@@ -69,4 +69,4 @@
 
     </Border>
 
-</Page>
+</local:SpecialErroBasePage>
